+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.Int.csproj
@@ -32,34 +32,34 @@
   </ItemGroup>
 
   <ItemGroup>       
-    <Compile Include="$(SharedSrcBaseDir)\Temporal.TestUtil\public\TestBase.cs"
-             Link="$(SharedSrcBaseLabel)\Temporal.TestUtil\public\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedSrcBaseDir)\Temporal.TestUtil\public\TestBase.cs" Link="$(SharedSrcBaseLabel)\Temporal.TestUtil\public\%(Filename)%(Extension)" />       
+    <Compile Include="..\..\Shared\Temporal.TestUtil\internal\TestTlsOptions.cs">
+      <Link>Shared-Test-Src\Temporal.TestUtil\internal\TestTlsOptions.cs</Link>
+    </Compile>       
+    <Compile Update="Temporal.Sdk.WorkflowClient.Test.Int\public\TemporalClientTlsTest.cs">
+      <Link>Shared-Test-Src\Temporal.TestUtil\internal\CertificateKeyPair.cs</Link>
+    </Compile>       
+    <Compile Update="Temporal.Sdk.WorkflowClient.Test.Int\public\TemporalClientMutualTlsTest.cs">
+      <Link>Shared-Test-Src\Temporal.TestUtil\internal\CertificateKeyPair.cs</Link>
+    </Compile>
   </ItemGroup>
   
   <ItemGroup>    
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ITemporalTestServerController.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ITemporalTestServerController.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
   
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\JavaBasedTemporalTestServerController.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\JavaBasedTemporalTestServerController.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TemporalLiteExeTestServerController.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TemporalLiteExeTestServerController.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\IntegrationTestBase.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\IntegrationTestBase.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TestEnvironment.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TestEnvironment.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ProcessManager.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ProcessManager.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ExtendedWorkflowServiceClient.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\ExtendedWorkflowServiceClient.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />
 
-    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TestCaseContextMonikers.cs"
-             Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />    
+    <Compile Include="$(SharedTestSrcBaseDir)\Temporal.TestUtil\internal\TestCaseContextMonikers.cs" Link="$(SharedTestSrcBaseLabel)\Temporal.TestUtil\internal\%(Filename)%(Extension)" />    
   </ItemGroup>
 
 </Project>
