+++ Src/Test/Shared/Temporal.TestUtil/internal/ITemporalTestServerController.cs
@@ -4,7 +4,7 @@ namespace Temporal.TestUtil
 {
     internal interface ITemporalTestServerController
     {
-        Task StartAsync();
+        Task StartAsync(TestTlsOptions testTlsOptions, int port = 7233);
         Task ShutdownAsync();
     }
 }
