+++ Src/Test/Shared/Temporal.TestUtil/internal/JavaBasedTemporalTestServerController.cs
@@ -5,7 +5,7 @@ namespace Temporal.TestUtil
 {
     internal class JavaBasedTemporalTestServerController : ITemporalTestServerController
     {
-        public Task StartAsync()
+        public Task StartAsync(TestTlsOptions tlsOptions, int port = 7233)
         {
             throw new NotImplementedException($"{nameof(JavaBasedTemporalTestServerController)} is not implemented.");
         }
