+++ Src/Test/Shared/Temporal.TestUtil/internal/TestTlsOptions.cs
@@ -1,0 +1,9 @@
+namespace Temporal.TestUtil
+{
+    public enum TestTlsOptions
+    {
+        None = 0,
+        Server = 1,
+        Mutual = 2
+    }
+}
